Reference NSRT test files from NonexistentOffsetInArrayDimFetchRuleTest instead of duplicating
